add confirm statement to delete

+++ helpers/deleteBranch.js
@@ -7,20 +7,22 @@ import { bouncingBall } from "cli-spinners";
 export default async () => {
   let spinner;
   try {
-    const { branch } = await createBranchPromptPromise({
+    const { branch, confirmed } = await createBranchPromptPromise({
       message: "Which branch would you like to delete?",
       shouldConfirm: true,
     });
 
-    if (branch) {
-      spinner = ora({
-        spinner: bouncingBall,
-        text: `Branch Lightyear - deleting branch: ${branch}`,
-      });
+    if (confirmed) {
+      if (branch) {
+        spinner = ora({
+          spinner: bouncingBall,
+          text: `Branch Lightyear - deleting branch: ${branch}`,
+        });
 
-      const deleteScript = `git branch -D ${branch}`;
-      const { stderr } = await exec(deleteScript);
-      spinner.succeed(stderr);
+        const deleteScript = `git branch -D ${branch}`;
+        const { stderr } = await exec(deleteScript);
+        spinner.succeed(stderr);
+      }
     }
   } catch (error) {
     spinner.fail(`Branch Lightyear - error while deleting branch`);
